requirements.txt
PySide6>=6.11,<6.12
pypdf>=6.9,<7.0
PyMuPDF>=1.27,<1.28
pdf2image>=1.17,<1.18
Pillow>=12.2,<13.0
pytesseract>=0.3.13,<0.4
reportlab>=4.4,<5.0
pdfplumber>=0.11,<0.12
python-docx>=1.2,<2.0
python-pptx>=1.0,<2.0
openpyxl>=3.1,<4.0
pdf2docx>=0.5,<0.6
pyinstaller>=6.19,<7.0
